refactor: 추천페이지 가로 무한스크롤 뷰쪽 코드 수정
MultiView.Horizontal이 검색 키워드를 들고 있어야 무한스크롤 할 때 어떤 키워드로 메서드를 호출할 건지 정할 수 있어서 추가해줌.
파라미터 추가된 거에 맞춰서 프래그먼트에서 멀티뷰 만들 때랑 뷰모델에서 RecommendationHorizontalData 만들 때에도 키워드 넣어줌.
