+++ app/src/main/java/com/wannabeinseoul/seoulpublicservice/ui/recommendation/RecommendationFragment.kt
@@ -9,7 +9,6 @@ import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.wannabeinseoul.seoulpublicservice.SeoulPublicServiceApplication
 import com.wannabeinseoul.seoulpublicservice.databinding.FragmentRecommendationBinding
 import com.wannabeinseoul.seoulpublicservice.ui.detail.DetailFragment
@@ -134,23 +133,17 @@ class RecommendationFragment : Fragment() {
     private val randomTip: String = tipsMap[randomTipHeader]?.random() ?: ""
 
     private fun initViewModel() = viewModel.let { vm ->
-        val onScrollListener = object : RecyclerView.OnScrollListener() {
-            override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
-                val lastVisiblePosition = (recyclerView.layoutManager as LinearLayoutManager)
-                    .findLastCompletelyVisibleItemPosition()
-                val lastPosition = recyclerView.adapter!!.itemCount - 1
+        // 프래그먼트에서 해줄 건 뷰모델 함수 호출하는 람다식(== 함수) 넘겨주는게 끝.
+//        fun fn1(query: String, num: Int) = vm.fn1(query, num)
+        fun fn1(query: String, num: Int) = listOf<RecommendationData>()
 
-                if (lastVisiblePosition == lastPosition) {
-                    // 뷰모델 함수 호출 (무한 스크롤 명령 - 데이터 더 꺼내기)
-                }
-            }
-        }
         vm.horizontalDataList.observe(viewLifecycleOwner) { horizontalDataList ->
             val multiViews: MutableList<RecommendationAdapter.MultiView> = horizontalDataList.map {
                 RecommendationAdapter.MultiView.Horizontal(
+                    it.keyword,
                     it.title,
                     RecommendationHorizontalAdapter(showDetailFragment).apply { submitList(it.list) },
-                    onScrollListener
+                    ::fn1
                 )
             }.toMutableList()
             if (multiViews.size >= 1) {
