+++ app/src/main/java/com/wannabeinseoul/seoulpublicservice/ui/recommendation/RecommendationHorizontalData.kt
@@ -1,6 +1,7 @@
 package com.wannabeinseoul.seoulpublicservice.ui.recommendation
 
 data class RecommendationHorizontalData(
+    val keyword: String,
     val title: String,
     val list: List<RecommendationData>,
 )
